Add import cluster API for adding existing clusters (#120)
This API is used to help users import their existing clusters into the Kvrocks controller,
the flow is like the create API but cluster node information comes from existing nodes. 
For now, we allow passing the existing node list but only parses it from the first one, 
maybe we can retry other nodes if the first one is not broken.